refactor modal to modular base component
